--- packages/react/src/components/adapter-context/index.ts
@@ -1,5 +1,0 @@
-export {
-  AdapterContext,
-  createAdapterContext,
-  selectAdapterConfigurationStatus,
-} from './adapter-context';
