--- packages/react/src/components/reconfigure-adapter/index.ts
@@ -1,1 +1,0 @@
-export { ReconfigureAdapter } from './reconfigure-adapter';
