--- packages/react/src/components/toggle-feature/index.ts
@@ -1,5 +1,0 @@
-import type { TProps as TToggleProps } from './toggle-feature';
-
-export type TProps = TToggleProps;
-
-export { ToggleFeature } from './toggle-feature';
