+++ packages/react/src/configure-adapter/configure-adapter.tsx
