+++ packages/react/src/configure-adapter/helpers.ts
