+++ packages/react/src/configure-adapter/index.ts
