+++ packages/react/src/get-is-feature-enabled.ts
@@ -6,7 +6,7 @@ import type {
 } from '@flopflip/types';
 
 import { DEFAULT_FLAG_PROP_KEY } from '../../constants';
-import { getFlagVariation } from '../get-flag-variation';
+import { getFlagVariation } from './get-flag-variation';
 
 const getIsFeatureEnabled = (
   allFlags: TFlagsContext,
