--- packages/react/src/helpers/get-flag-variation/index.ts
@@ -1,1 +1,0 @@
-export { getFlagVariation } from './get-flag-variation';
