--- packages/react/src/helpers/get-is-feature-enabled/index.ts
@@ -1,1 +1,0 @@
-export { getIsFeatureEnabled } from './get-is-feature-enabled';
