--- packages/react/src/helpers/get-normalized-flag-name/index.ts
@@ -1,1 +1,0 @@
-export { getNormalizedFlagName } from './get-normalized-flag-name';
