--- packages/react/src/helpers/is-nil/index.ts
@@ -1,1 +1,0 @@
-export { isNil } from './is-nil';
