--- packages/react/src/hocs/index.ts
@@ -1,2 +1,0 @@
-export { setDisplayName } from './set-display-name';
-export { wrapDisplayName } from './wrap-display-name';
