--- packages/react/src/hocs/set-display-name/index.ts
@@ -1,1 +1,0 @@
-export { setDisplayName } from './set-display-name';
