--- packages/react/src/hocs/wrap-display-name/index.ts
@@ -1,1 +1,0 @@
-export { wrapDisplayName } from './wrap-display-name';
