--- packages/react/src/hooks/use-adapter-context/index.ts
@@ -1,1 +1,0 @@
-export { useAdapterContext } from './use-adapter-context';
