--- packages/react/src/hooks/use-adapter-reconfiguration/index.ts
@@ -1,1 +1,0 @@
-export { useAdapterReconfiguration } from './use-adapter-reconfiguration';
