--- packages/react/src/hooks/use-adapter-subscription/index.ts
@@ -1,1 +1,0 @@
-export { useAdapterSubscription } from './use-adapter-subscription';
