+++ packages/react/src/index.ts
@@ -1,4 +1,4 @@
-import type { TProps as ToggleFeatureProps } from './components/toggle-feature';
+import type { TProps as ToggleFeatureProps } from './toggle-feature';
 
 export type TToggleFeatureProps = ToggleFeatureProps;
 
@@ -8,21 +8,22 @@ export {
   AdapterContext,
   ConfigureAdapter,
   createAdapterContext,
-  ReconfigureAdapter,
   selectAdapterConfigurationStatus,
-  ToggleFeature,
-} from './components';
+} from './adapter-context';
+export { ReconfigureAdapter } from './reconfigure-adapter';
+export { ToggleFeature } from './toggle-feature';
 export {
   ALL_FLAGS_PROP_KEY,
   DEFAULT_FLAG_PROP_KEY,
   DEFAULT_FLAGS_PROP_KEY,
 } from './constants';
-export { getFlagVariation, getIsFeatureEnabled, isNil } from './helpers';
-export { setDisplayName, wrapDisplayName } from './hocs';
-export {
-  useAdapterContext,
-  useAdapterReconfiguration,
-  useAdapterSubscription,
-} from './hooks';
+export { getFlagVariation } from './/get-flag-variation';
+export { getIsFeatureEnabled } from './get-is-feature-enabled';
+export { isNil } from './is-nil';
+export { wrapDisplayName } from './wrap-display-name';
+export { setDisplayName } from './set-display-name';
+export { useAdapterContext } from './use-adapter-context';
+export { useAdapterReconfiguration } from './use-adapter-reconfiguration';
+export { useAdapterSubscription } from './use-adapter-subscription';
 
 export { version };
