+++ packages/react/src/reconfigure-adapter.ts
