+++ packages/react/src/use-adapter-reconfiguration.ts
