+++ packages/react/src/use-adapter-subscription.ts
