+++ packages/react/test/configure-adapter/helpers.spec.js
