+++ packages/react/test/get-is-feature-enabled.spec.js
@@ -1,5 +1,5 @@
 import { describe, expect, it, vi } from 'vitest';
-import { getIsFeatureEnabled } from './get-is-feature-enabled';
+import { getIsFeatureEnabled } from '../src/get-is-feature-enabled';
 
 vi.mock('tiny-warning');
