+++ packages/react/test/get-normalized-flag-name.spec.js
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { getNormalizedFlagName } from './get-normalized-flag-name';
+import { getNormalizedFlagName } from '../src/get-normalized-flag-name';
 
 describe('when not camel caased', () => {
   it('should normalized the flag name', () => {
