+++ packages/react/test/wrap-display-name.spec.js
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { wrapDisplayName } from './wrap-display-name';
+import { wrapDisplayName } from '../src/wrap-display-name';
 
 function BaseComponent() {
   return 'BaseComponent';
